prepare e2e testing
